Adding EPUB
I've put EPUB in new standards for now (since it's technically "new" to W3C), but if someone feels it is better placed elsewhere, please feel free to move it. I've also added our TF work on Fixed Layout with the documenting additional guidance section, as that is related to the goals for that document. 

Also fixed a small typo I caught in the description for the Low Vision guidance.